Working on code touchups
integrated namespaces more, will make a state machine class
