+++ Assets/Scripts/Core/MonoSingleton.cs.meta
