+++ Assets/Scripts/Core/StateMachine/State.cs
@@ -1,0 +1,9 @@
+namespace Core.StateMachine
+{
+    public abstract class State
+    {
+        public virtual void Enter() { }
+        public virtual void Exit() { }
+        public virtual void Update() { }
+    }
+}
