+++ Assets/Scripts/Managers/DiscordManager.cs
@@ -1,16 +1,11 @@
-using UnityEngine;
 
 // TODO: Implement Discord Rich Presence integration
-public class DiscordManager : Singleton<DiscordManager>
+
+using Core;
+
+namespace Managers
 {
-    // Start is called once before the first execution of Update after the MonoBehaviour is created
-    void Start()
-    {
-
-    }
-
-    // Update is called once per frame
-    void Update()
+    public class DiscordManager : Singleton<DiscordManager>
     {
 
     }
