+++ Assets/Scripts/MouseLook.cs
@@ -1,3 +1,4 @@
+using Managers;
 using UnityEngine;
 using UnityEngine.InputSystem;
