+++ Assets/Scripts/Player/States/PlayerState.cs
@@ -1,3 +1,5 @@
+using Core;
+
 namespace Player.States
 {
     public abstract class PlayerState : State
