--- Assets/Scripts/State.cs
@@ -1,6 +1,0 @@
-public abstract class State
-{
-    public virtual void Enter() { }
-    public virtual void Exit() { }
-    public virtual void Update() { }
-}
